Establish a first-pass at the SafetyNet architecture
This commit adds the files and function calls that I expect we'll need
for the project. All this is still completely up for debate, but
even having these stubs should help us communicate about dependencies
and progress.

I've also added in some code for parsing the Okutama
SingleActionTrackingLabels data annotations, which look like they support
the pedestrian persistence that we're looking for. This should help
me start development before we have the pedestrian segmentation working.